+++ backend/tests/test_report_filters.py
@@ -9,7 +9,7 @@ import pytest
 from httpx import AsyncClient
 from datetime import date
 
-from app.routers.reports import (
+from app.routers.report_helpers import (
     filter_transactions_by_accounts,
     filter_transactions_by_merchants,
     parse_filter_param,
